v2 - remove libraw and aces_container (#167)
Changes:

replaces direct Libraw calls with OIIO for reading raw images
replaces AcesContainer with OIIO for writing exr images
replaces all pixel operations with OIIO IBAs
Since not all Libraw flags are exposed in the raw input plugin of OIIO, these command line parameters have been removed:
-P - bad pixel mask
-K - dark frame
-j - fuji-rotate
-m - median filter
-f - four-colour RGB
-T - print Libraw-supported cameras
-F - use big file
-E - mmap-ed IO
-s - image index in the file

We may want to reintroduce them on the OIIO side later, if required.

This is a breaking change and not supposed to be compatible with v1.x.

________________


Signed-off-by: Anton Dukhovnikov <antond@wetafx.co.nz>
Co-authored-by: Aleksandr Motsjonov <soswow@users.noreply.github.com>